+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Dependency: TypeScript
 - `LexOrder` class
 - `LexOrderOptions` interface
+- `SymbolConverter` interface
 
 
 [unreleased]: https://github.com/aryelgois/lexorder/compare/initial-commit...HEAD